refactor: implement handler for adding a player's ships (message "add_ships")
